+++ src/utils/read.ts
@@ -1,9 +1,8 @@
 import { join } from 'node:path';
 import { logger } from './logger';
 import { config } from 'dotenv';
-import type { BaseDeployOptions, Environment, EsbuildAlias } from '$types';
+import type { BaseDeployOptions, Environment } from '$types';
 import { getEnvironmentFileName } from './common';
-import { readFile } from 'node:fs/promises';
 import { execute } from './execute';
 
 export const validateProject = async ({
@@ -32,33 +31,6 @@ export const validateProject = async ({
 		cwd: projectRoot,
 		commandArguments,
 	});
-};
-
-export const getAlias = async (options: {
-	tsconfig?: string;
-	projectRoot: string;
-	workspaceRoot: string;
-}) => {
-	const { projectRoot, workspaceRoot } = options;
-	let alias = await getEsbuildAliasFromTsConfig(
-		projectRoot,
-		options.tsconfig,
-	);
-	if (!alias) {
-		if (options.tsconfig && options.tsconfig !== 'tsconfig.json') {
-			alias = await getEsbuildAliasFromTsConfig(
-				projectRoot,
-				'tsconfig.json',
-			);
-		}
-		if (!alias) {
-			alias = await getEsbuildAliasFromTsConfig(
-				workspaceRoot,
-				'tsconfig.base.json',
-			);
-		}
-	}
-	return alias;
 };
 
 export const getEnvironment = async (options: {
@@ -138,76 +110,6 @@ export const findContentBetween = (
 	return string.substring(start + 1, end).trim();
 };
 
-/**
- * Get the aliases from the base tsconfig.json file in the monorepo.
- *
- * @example
- *
- * ```json
- * // base.tsconfig.json in the workspace root:
- * "compilerOptions": {
- * 	"baseurl": ".",
- * 	"paths": {
- * 		"@shared/utils":["libs/shared/utils/src"],
- * 		"@shared/types": ["libs/shared/types/src"]
- * 	}
- * }
- * ...
- * ```
- *
- * ```ts
- * const aliases = getAliases('workspace-root');
- *
- * // aliases = {
- * // 	'@shared/utils': '@shared/utils/src/index.ts',
- * // 	'@shared/types': '@shared/types/src/index.ts',
- * // }
- * ```
- *
- * @param root The absolute path to the root
- * @param baseTsConfigFileName the name of the base tsconfig file. Defaults to
- *   `tsconfig.json`
- * @returns the aliases from the base tsconfig file
- */
-const getEsbuildAliasFromTsConfig = async (
-	root: string,
-	baseTsConfigFileName = 'tsconfig.json',
-): Promise<EsbuildAlias | undefined> => {
-	try {
-		const baseTsConfig = await readFile(
-			join(root, baseTsConfigFileName),
-			'utf8',
-		);
-		const config = JSON.parse(removeComments(baseTsConfig));
-		const paths = config?.compilerOptions?.paths;
-		if (!paths) {
-			return;
-		}
-
-		const alias: EsbuildAlias = {};
-
-		const fixPath = (key: string, path: string) => {
-			if (key.endsWith('/*') || path.endsWith('.ts')) {
-				return join(root, path);
-			}
-
-			return join(root, path, 'index.ts');
-		};
-
-		for (const [key, value] of Object.entries(paths)) {
-			if (Array.isArray(value) && typeof value[0] === 'string') {
-				alias[key] = fixPath(key, value[0]);
-			} else if (typeof value === 'string') {
-				alias[key] = fixPath(key, value);
-			}
-		}
-		return alias;
-	} catch (error) {
-		logger.error('getEsbuildAliasFromTsConfig', error);
-		return;
-	}
-};
-
 const setEnvironment = (line: string, environment: Environment): void => {
 	// get all instances of `=` in the line
 	const equalSignIndexes = getIndexesOf(line, '=');
@@ -273,25 +175,4 @@ const getSecondToFirstIndexOf = (string: string, character: string): number => {
 		return -1;
 	}
 	return string.indexOf(character, index + 1);
-};
-
-/**
- * Remove all comments from a string of code.
- *
- * @example
- *
- * ```ts
- * const code = `
- * // This is a comment
- * const a = 1;
- * `;
- *
- * removeComments(code); // 'const a = 1;'
- * ```
- *
- * @param code the string of code to remove comments from
- * @returns the string of code without comments
- */
-const removeComments = (code: string): string => {
-	return code.replace(/\/\*[\S\s]*?\*\/|\/\/.*/g, '').trim();
 };